+++ fragments/labels/macadminspython.sh
@@ -1,0 +1,9 @@
+macadminspython)
+    name="MacAdmins Python"
+    type="pkg"
+    packageID="org.macadmins.python.recommended"
+    downloadURL=$(curl --silent --fail "https://api.github.com/repos/macadmins/python/releases/latest" | awk -F '"' "/browser_download_url/ && /python_recommended_signed/ { print \$4; exit }")
+    appNewVersion=$(grep -o -E '\d+\.\d+\.\d+\.\d+' <<< $downloadURL | head -n 1)
+    expectedTeamID="9GQZ7KUFR6"
+    blockingProcesses=( NONE )
+    ;;